Proposal for fixing issue #1784 (Spice throws error at midline semicolon comments)
